Add link to MTurk handler
Previously the "/api/turk" route would create or login a turker and redirect them to the url in the redirectURL parameter.
This caused the tagging experiment to be rendered in an iframe on the MTurk website (which we don't want - too many Amazon buttons and ui).
This changes renders a link that opens in a new tab by default to take the turker out of the MTurk environment.